requirements.txt
# Install system packages first (Raspberry Pi OS):
# sudo apt install python3-picamera2 python3-opencv python3-numpy python3-psutil

fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pyyaml>=6.0
psutil>=5.9.0
numpy>=1.24.0
onnxruntime>=1.16.0
ultralytics>=8.0.0
onnx>=1.12.0,<2.0.0
onnxslim>=0.1.71
